services/ingestion/requirements.txt
kafka-python
sseclient-py
requests